update(cmd): revamp `delete falco` command
